fix: load router by middleware
